👨‍🔬 Add mix task & remove scripts
