+++ lib/code_vis/function_tracer.ex
@@ -1,4 +1,4 @@
-defmodule TreeTracer do
+defmodule CodeVis.FunctionTracer do
   @moduledoc """
   Trace remote function calls, aggregating them into an ets table
   Key: caller_mfa
@@ -13,55 +13,15 @@ defmodule TreeTracer do
     ]
 
   Ideally at the end we can recursively build a tree from a top level function of all the dependants
+
+  End result
+  ETS table `:functions` with an entry for every Function called and a list of all of the functions called from within that function
   """
 
-  def run() do
-    :ets.new(:functions, [:named_table, :public, :duplicate_bag])
-    Mix.Task.clear()
-    Mix.Task.run("compile", ["--force", "--tracer", __MODULE__])
-
-    IO.puts("Trace complete. Building Tree…")
-
-    root = {CodeVis, :i_alias, 0}
-    # root = {Display, :as_file, 2}
-
-    # tree = build_tree(root)
-    map = build_map(root)
-
-    # callees = get_callees(root)
-
-    IO.puts("Results for: #{Display.format_mfa(root)}\n")
-    # IO.puts("#{Display.format_mfa(root)} -> #{format_ets_result(callees)}")
-    # IO.inspect(map)
-    Display.as_io(map, root)
-    Display.as_file(map, root)
-  end
-
-  @spec get_remote_calls(mfa()) :: [mfa()]
-  def get_remote_calls(mfa) do
-    :functions
-    |> :ets.lookup(mfa)
-    |> Enum.map(&elem(&1, 1))
-  end
-
-  @doc """
-  Alternate implementation, create a flat map and recurse over it (basically just like in the ets table)
-  """
-  @spec build_map(map(), mfa()) :: map()
-  def build_map(accumulator \\ %{}, current_mfa) do
-    case get_remote_calls(current_mfa) do
-      [] ->
-        # base case
-        Map.put(accumulator, current_mfa, [])
-
-      remote_mfas ->
-        updated = Map.put(accumulator, current_mfa, remote_mfas)
-
-        Enum.reduce(remote_mfas, updated, fn next_mfa, acc ->
-          build_map(acc, next_mfa)
-        end)
-    end
-  end
+  # def trace({:imported_function, meta, module, name, arity}, env) do
+  #   Import2Alias.Server.record(env.file, meta[:line], meta[:column], module, name, arity)
+  #   :ok
+  # end
 
   # 1. Add more context to what's happening in trace/2
   # 2. What about a Module white list rather than black list?
@@ -97,5 +57,3 @@ defmodule TreeTracer do
 
   def trace(_, _), do: :ok
 end
-
-TreeTracer.run()
